ISSUE #4359 - move selects to parent level
