+++ frontend/src/v5/ui/routes/dashboard/projects/tickets/selectMenus/selectMenus.styles.ts
